feat: support bundleDependencies: true
What ends up in the parsed output is still an array of names, but we
generate that name from the keys if it's set to true.

Also, avoid the footgun of people not realizing that bundleDependencies
should be an array of names, rather than a {name:version} object.

If bundleDependencies is something other than a boolean or array, it is
removed from the manifest.